finalized the cos-to-sql sample code and readme instructions
