+++ cos-to-sql/.dockerignore
@@ -1,0 +1,5 @@
+.dockerignore
+.gitignore
+build
+Dockerfile
+node_modules
