+++ cos-to-sql/Dockerfile
@@ -1,21 +1,21 @@
-FROM registry.access.redhat.com/ubi9/nodejs-22:latest AS build-env
+# Download dependencies in builder stage
+FROM registry.access.redhat.com/ubi9/nodejs-22:latest AS builder
+
+COPY --chown=${CNB_USER_ID}:${CNB_GROUP_ID} package.json package-lock.json /app/
 WORKDIR /app
+RUN npm ci --omit=dev
 
-# Copy job dependency manifests to the container image.
-COPY package.json ./
-
-# Install production dependencies.
-RUN npm install --production
 
 # Use a small distroless image for as runtime image
-FROM gcr.io/distroless/nodejs22-debian12
+FROM gcr.io/distroless/nodejs22
 
+COPY --chown=1001:0 --from=builder /app/node_modules /app/node_modules
+COPY --chown=1001:0 app.mjs /app
+COPY --chown=1001:0 utils/ /app/utils
+COPY --chown=1001:0 public/ /app/public
+
+USER 1001:0
 WORKDIR /app
-
-# Copy local code to the container image.
-COPY ./src ./
-
-# Copy the dependencies and other project related files
-COPY --from=build-env /app ./
 EXPOSE 8080
+
 CMD ["app.mjs"]
